fix(user): 修改 DepartmentRepository 中 findByDepartmentId 参数类型
将 findByDepartmentId 方法的参数类型从 int 修改为 long，以确保与数据库字段类型一致，避免潜在的数据转换问题。

+++ src/main/java/top/xcyyds/wxfbackendclient/module/user/persistence/repository/DepartmentRepository.java
@@ -13,6 +13,6 @@ import top.xcyyds.wxfbackendclient.module.user.pojo.entity.Department;
 public interface DepartmentRepository extends JpaRepository<Department,Long> {
 
 
-    Department findByDepartmentId(int departmentId);
+    Department findByDepartmentId(long departmentId);
 }
